Apply bounds at useDrag
