+++ components/player/SpeechBubble.tsx
@@ -54,24 +54,12 @@ const Bubble = styled<'div', BubbleProps>('div')`
 `
 
 export const SpeechBubble = ({ content, time, kind }: Props) => {
-  const [{ x, y }, set] = useSpring(() => ({ x: 0, y: 0 }))
-
-  const bind = useDrag(({ down, movement: [mx, my] }) => {
-    set({ x: down ? mx : 0, y: down ? my : 0 })
-  })
-
   return (
-    <animated.div {...bind()} style={
-      {
-        transform: x.interpolate(x => `translateX(${x}px)`),
-      }
-    } >
-      <Bubble
-        color={COMMENT_COLOR[kind]}
-        time={time}
-      >
-        {content}
-      </Bubble>
-    </animated.div>
+    <Bubble
+      color={COMMENT_COLOR[kind]}
+      time={time}
+    >
+      {content}
+    </Bubble>
   )
 }
